+++ testproject/settings.py
@@ -50,7 +50,7 @@ MIDDLEWARE = [
     "django.contrib.auth.middleware.AuthenticationMiddleware",
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
-    "books.middleware.TestMiddleware",
+    "inspector.middleware.InspectorMiddleware",
 ]
 
 ROOT_URLCONF = "testproject.urls"
